+++ modules/ui/sections/raw_tag_editor.js
@@ -11,6 +11,7 @@ import { t } from '../../core/localizer';
 import { utilArrayDifference, utilArrayIdentical } from '../../util/array';
 import { utilGetSetValue, utilNoAuto, utilRebind, utilTagDiff } from '../../util';
 
+
 export function uiSectionRawTagEditor(id, context) {
 
     var section = uiSection(id, context)